+++ resources/views/maintenance/ppe-item/index.blade.php
@@ -57,17 +57,27 @@
 								</div>
 						</div>
 						<div class="btn-group">
-							<button class="btn default dropdown-toggle" data-toggle="dropdown">Actions <i class="fa fa-angle-down"></i>
-							</button>
-							<ul class="dropdown-menu">
-								<li>
-									<a href="#" onclick="delete_items()"><i class="fa fa-trash-o"></i> Delete</a>
-								</li>
-							</ul>
+							@if($delete)
+								<button class="btn default dropdown-toggle" data-toggle="dropdown">Actions <i class="fa fa-angle-down"></i>
+								</button>
+								<ul class="dropdown-menu">
+									<li>
+										<a href="#" onclick="delete_items()"><i class="fa fa-trash-o"></i> Delete</a>
+									</li>
+								</ul>
+							@else
+								<button disabled class="btn default dropdown-toggle" data-toggle="dropdown">Actions <i class="fa fa-angle-down"></i>
+								</button>
+							@endif
 						</div>
 
 						<div class="btn-group pull-right" style="margin-left: 5px;">
-							<button class="btn blue" data-toggle="modal" href="#add-ppe-item">Add New <i class="fa fa-plus"></i></button>
+						@if($create)
+								<button class="btn blue" data-toggle="modal" href="#add-ppe-item">Add New <i class="fa fa-plus"></i></button>
+							@else
+								<button disabled class="btn blue" data-toggle="modal" href="#add-category">Add New <i class="fa fa-plus"></i></button>
+							@endif
+							<!-- <button class="btn blue" data-toggle="modal" href="#add-ppe-item">Add New <i class="fa fa-plus"></i></button> -->
 						</div>
 						<div class="btn-group pull-right">
 							<form id="search_form" class="form-inline">
@@ -125,8 +135,18 @@
 											{{ $item->addedBy }}
 										</td>
 										<td>
-											<a href="#" class="btn default btn-xs blue" onclick="update_item('{{$item->id}}','{{$item->main}}','{{$item->type}}')">Edit </a>
-											<a href="#" class="btn default btn-xs red" onclick="item_delete('{{$item->id}}')">Delete </a>
+											@if($edit)
+												<a href="#" class="btn default btn-xs blue" onclick="update_item('{{$item->id}}','{{$item->main}}','{{$item->type}}')">Edit </a>
+											@else
+												<button disabled href="#" class="btn default btn-xs blue" >Edit </button>
+											@endif
+											@if($delete)
+												<a href="#" class="btn default btn-xs red" onclick="item_delete('{{$item->id}}')">Delete </a>
+											@else
+												<button disabled href="#" class="btn default btn-xs red" >Delete </button>
+											@endif
+											<!-- <a href="#" class="btn default btn-xs blue" onclick="update_item('{{$item->id}}','{{$item->main}}','{{$item->type}}')">Edit </a>
+											<a href="#" class="btn default btn-xs red" onclick="item_delete('{{$item->id}}')">Delete </a> -->
 										</td>
 									</tr>
 								@empty
